Make scopes required
When calling the commercetools API you need to provide the same or a subset of scopes that have been provided to the API client. I.e. you can't request manage_extensions and manage_types when the client has been given the manage_project scope.

As a result we're going to require the scope so it is explicit